% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_create_blocks.R
\docType{data}
\name{lucC_blocks_raster_merge}
\alias{lucC_blocks_raster_merge}
\title{Merge blocks of RasterBrick or Rasters in accordance with number of blocks}
\usage{
lucC_blocks_raster_merge (path_open_GeoTIFFs = NULL,
pattern_name = NULL, is.rasterBrick = FALSE)
}
\arguments{
\item{path_open_GeoTIFFs}{Character. Name a path folder to OPEN raster images data.}

\item{pattern_name}{Character. A pattern in name of GeoTIFF to mosaic them}

\item{is.rasterBrick}{Boolean. If TRUE GeoTIFF is a RasterBrick, FALSE is a single layer. Default is FALSE}
}
\value{
RasterBrick Mosaic.
}
\description{
Merge GeoTIFF splitted into parts. \url{https://stackoverflow.com/questions/29784829/r-raster-package-split-image-into-multiples}
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterItanhanga.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)

# blocks saved in folder
lucC_blocks_raster_create(raster_obj = rb_class, number_blocks_xy = 2, save_images = TRUE)

lucC_blocks_raster_merge(path_open_GeoTIFFs = paste0(getwd(), "/Blocks_RasterBrick", sep = ""),
                         pattern_name = "Raster_Block_", is.rasterBrick = TRUE)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
