% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_create_RasterBrick.R
\docType{data}
\name{lucC_create_RasterBrick}
\alias{lucC_create_RasterBrick}
\title{Save tibble as Raster in Folder}
\usage{
lucC_create_RasterBrick(path_open_GeoTIFFs = NULL, path_save_RasterBrick = NULL)
}
\arguments{
\item{path_open_GeoTIFFs}{Character. Name a path folder to OPEN raster images data.}

\item{path_save_RasterBrick}{Character. Name a path folder to SAVE RasterBrick from GeoTIFF images. If  doesn't exist, a new directory is created}
}
\value{
RasterBrick file
}
\description{
Create a RasterBrick from list of GeoTIFF images stored in a folder
}
\examples{
\dontrun{
library(lucCalculus)

# create a RasterBrick from individual raster GeoTIFF classified previously
lucC_create_RasterBrick(path_open_GeoTIFFs =
                        c(system.file("extdata/raster/rasterSample", package = "lucCalculus")),
                        path_save_RasterBrick = getwd())

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
