% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations_extra.r
\docType{data}
\name{lucC_merge}
\alias{lucC_merge}
\title{LucC Merge}
\usage{
lucC_merge(first_raster = NULL, second_raster = NULL)
}
\arguments{
\item{first_raster}{Matrix. An interval between two dates.}

\item{second_raster}{Matrix. An interval between two dates.}
}
\value{
Data set merged with two data sets that values are in the same interval
}
\description{
Merge two data sets by different columns and rows. Because base::merge replace values in initial years
}
\examples{
\dontrun{
library(lucCalculus)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
