% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_update_raster.R
\docType{data}
\name{lucC_raster_update}
\alias{lucC_raster_update}
\title{Update a RasterBrick with pixel replaced}
\usage{
lucC_raster_update(raster_obj = NULL, data_mtx = NULL,
timeline = NULL, class_to_replace = NULL, new_pixel_value = 20)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{timeline}{Character. A list of all dates of classified raster, timeline}

\item{class_to_replace}{Character. All labels of each value of pixel from classified raster}

\item{new_pixel_value}{Integer. New pixel value to raster. Default is 20}
}
\value{
Matrix with raster and new pixel to create a RasterBrick reclassified
}
\description{
Update a RasterBrick with new values of pixel discovered from LUC Calculus formalism
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterSample.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)
my_label <- c("Degradation", "Fallow_Cotton", "Forest", "Pasture", "Soy_Corn", "Soy_Cotton",
              "Soy_Fallow", "Soy_Millet", "Soy_Sunflower", "Sugarcane", "Urban_Area", "Water")
my_timeline <- c("2001-09-01", "2002-09-01", "2003-09-01", "2004-09-01", "2005-09-01",
                 "2006-09-01", "2007-09-01", "2008-09-01", "2009-09-01", "2010-09-01",
                 "2011-09-01", "2012-09-01", "2013-09-01", "2014-09-01", "2015-09-01",
                 "2016-09-01")

a <- lucC_pred_recur(raster_obj = rb_class, raster_class = "Forest",
                     time_interval1 = c("2001-09-01","2001-09-01"),
                     time_interval2 = c("2002-09-01","2016-09-01"),
                     label = my_label, timeline = my_timeline)

# update original RasterBrick with new class
num_label <- length(my_label) + 1
rb_class_new <- lucC_raster_update(raster_obj = rb_class,
                                   data_mtx = a,
                                   timeline = my_timeline,
                                   class_to_replace = "Forest",  # the same class previously
                                   new_pixel_value = num_label)  # new pixel value

lucC_plot_bar_events(data_mtx = rb_class_new, pixel_resolution = 232, custom_palette = FALSE)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
