% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_allen_relations.r
\docType{data}
\name{lucC_relation_met_by}
\alias{lucC_relation_met_by}
\title{Allen Relation Met By}
\usage{
lucC_relation_met_by(first_raster = NULL, second_raster = NULL)
}
\arguments{
\item{first_raster}{Matrix. An interval between two dates.}

\item{second_raster}{Matrix. An interval between two dates.}
}
\value{
Data set with merge of two data sets
}
\description{
Provide an Allen's interval relation MET_BY which end time interval of the
raster_2 must be (=) equal to the start time interval from raster_1, and also time interval
from raster_2 can not overlap time interval from raster_1. See more at (ALLEN, J. F.
"Maintaining knowledge about temporal intervals". Communications of the ACM, v(26), 11,
1983, 832-843. DOI: \url{http://dx.doi.org/10.1145/182.358434})
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterSample.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)
my_label <- c("Degradation", "Fallow_Cotton", "Forest", "Pasture", "Soy_Corn", "Soy_Cotton",
              "Soy_Fallow", "Soy_Millet", "Soy_Sunflower", "Sugarcane", "Urban_Area", "Water")
my_timeline <- c("2001-09-01", "2002-09-01", "2003-09-01", "2004-09-01", "2005-09-01",
                 "2006-09-01", "2007-09-01", "2008-09-01", "2009-09-01", "2010-09-01",
                 "2011-09-01", "2012-09-01", "2013-09-01", "2014-09-01", "2015-09-01",
                 "2016-09-01")

lucC_plot_raster(raster_obj = rb_class, timeline = my_timeline, label = my_label,
                 custom_palette = FALSE, plot_ncol = 4)

a <- lucC_pred_holds(raster_obj = rb_class, raster_class = "Degradation",
                     time_interval = c("2004-09-01","2007-09-01"),
                     relation_interval = "contains", label = my_label,
                     timeline = my_timeline)

b <- lucC_pred_holds(raster_obj = rb_class, raster_class = "Forest",
                     time_interval = c("2001-09-01","2003-09-01"),
                     relation_interval = "contains", label = my_label,
                     timeline = my_timeline)

# met by
c <- lucC_relation_met_by(first_raster = a, second_raster = b)

lucC_plot_raster_result(raster_obj = rb_class, data_mtx = c, timeline = my_timeline,
                        label = my_label, custom_palette = FALSE, plot_ncol = 4)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
