% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_create_RasterBrick.R
\docType{data}
\name{lucC_remove_TmpFilesRaster}
\alias{lucC_remove_TmpFilesRaster}
\title{Remove Temporary files in Linux}
\usage{
lucC_remove_TmpFilesRaster()
}
\description{
Remove temporary files created by raster package
}
\examples{
\dontrun{
library(lucCalculus)

# create a RasterBrick from individual raster GeoTIFF classified previously
lucC_create_RasterBrick(path_open_GeoTIFFs =
                        c(system.file("extdata/raster/rasterSample", package = "lucCalculus")),
                        path_save_RasterBrick = getwd())
lucC_remove_TmpFilesRaster()

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
