% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_save_raster_GeoTIFF.R
\docType{data}
\name{lucC_save_GeoTIFF}
\alias{lucC_save_GeoTIFF}
\title{Save a RasterBrick reclassified in Folder}
\usage{
lucC_save_GeoTIFF(raster_obj = NULL, data_mtx = NULL,
path_raster_folder = NULL, as_RasterBrick = FALSE)
}
\arguments{
\item{raster_obj}{Raster. A raster brick with classified images}

\item{data_mtx}{Matrix. A matrix with pixel replaced, can be obtined using lucC_update_raster()}

\item{path_raster_folder}{Character. Name a path folder to save raster images data. If  doesn't exist, a new directory is created}

\item{as_RasterBrick}{Boolean. If FALSE, each layer will be saved in separate file. If TRUE, write a RasterBrick in a file. Default is FALSE.}
}
\value{
Raster in geotiff format to open using SIG
}
\description{
Save new_raster reclassified in a diretory defined by user.
NOTE: Applied only when update a raster with new values, after had applied the
function lucC_raster_update.
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterSample.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)
my_label <- c("Degradation", "Fallow_Cotton", "Forest", "Pasture", "Soy_Corn", "Soy_Cotton",
              "Soy_Fallow", "Soy_Millet", "Soy_Sunflower", "Sugarcane", "Urban_Area", "Water")
my_timeline <- c("2001-09-01", "2002-09-01", "2003-09-01", "2004-09-01", "2005-09-01",
                 "2006-09-01", "2007-09-01", "2008-09-01", "2009-09-01", "2010-09-01",
                 "2011-09-01", "2012-09-01", "2013-09-01", "2014-09-01", "2015-09-01",
                 "2016-09-01")

a <- lucC_pred_recur(raster_obj = rb_class, raster_class = "Forest",
                     time_interval1 = c("2001-09-01","2001-09-01"),
                     time_interval2 = c("2002-09-01","2016-09-01"),
                     label = my_label, timeline = my_timeline)

# update original RasterBrick with new class
num_label <- length(my_label) + 1
rb_class_new <- lucC_raster_update(raster_obj = rb_class,
                                   data_mtx = a,
                                   timeline = my_timeline,
                                   class_to_replace = "Forest",  # the same class previously
                                   new_pixel_value = num_label)  # new pixel value

lucC_plot_bar_events(data_mtx = rb_class_new, pixel_resolution = 232, custom_palette = FALSE)

# save the update matrix as GeoTIFF RasterBrick
lucC_save_GeoTIFF(raster_obj = rb_class, data_mtx = rb_class_new,
                  path_raster_folder = paste0(getwd(), "/updatedRasterRECUR", sep = ""),
                  as_RasterBrick = FALSE )

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
