% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_save_raster_GeoTIFF.R
\docType{data}
\name{lucC_save_raster_result}
\alias{lucC_save_raster_result}
\title{Update a RasterBrick with result from predicates}
\usage{
lucC_save_raster_result(raster_obj = NULL, data_mtx = NULL,
timeline = NULL, label = NULL, path_raster_folder = NULL, as_RasterBrick = FALSE)
}
\arguments{
\item{raster_obj}{Raster. A raster brick with classified images}

\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{timeline}{Character. A list of all dates of classified raster, timeline}

\item{label}{Character Vector. All labels of each value of pixel from classified raster}

\item{path_raster_folder}{Character. Name a path folder to save raster images data. If  doesn't exist, a new directory is created}

\item{as_RasterBrick}{Boolean. If FALSE, each layer will be saved in separate file. If TRUE, write a RasterBrick in a file. Default is FALSE.}
}
\value{
Matrix with raster and pixels to create a RasterBrick with result
}
\description{
Update a RasterBrick with new values of pixel discovered from LUC Calculus formalism to create GeoTIFF files. #' NOTE: Used after had applied some of the functions with predicate.
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterSample.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)
my_label <- c("Degradation", "Fallow_Cotton", "Forest", "Pasture", "Soy_Corn", "Soy_Cotton",
              "Soy_Fallow", "Soy_Millet", "Soy_Sunflower", "Sugarcane", "Urban_Area", "Water")
my_timeline <- c("2001-09-01", "2002-09-01", "2003-09-01", "2004-09-01", "2005-09-01",
                 "2006-09-01", "2007-09-01", "2008-09-01", "2009-09-01", "2010-09-01",
                 "2011-09-01", "2012-09-01", "2013-09-01", "2014-09-01", "2015-09-01",
                 "2016-09-01")

a <- lucC_pred_recur(raster_obj = rb_class, raster_class = "Forest",
                     time_interval1 = c("2001-09-01","2001-09-01"),
                     time_interval2 = c("2002-09-01","2016-09-01"),
                     label = my_label, timeline = my_timeline)

lucC_plot_raster_result(raster_obj = rb_class, data_mtx = a, timeline = my_timeline,
                        label = my_label, custom_palette = FALSE)
# save in file
lucC_save_raster_result(raster_obj = rb_class, data_mtx = a,
                        timeline = my_timeline, label = my_label,
                        path_raster_folder = paste0(getwd(), "/onlyRECUR", sep = ""),
                        as_RasterBrick = FALSE)

}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
