% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_save_raster_GeoTIFF.R
\docType{data}
\name{lucC_save_rasterBrick_by_layers}
\alias{lucC_save_rasterBrick_by_layers}
\title{Save a RasterBrick by layers}
\usage{
lucC_save_rasterBrick_by_layers(path_name_GeoTIFF_Brick = NULL)
}
\arguments{
\item{path_name_GeoTIFF_Brick}{Character. Name file and path folder to SPLIT rasterBrick by layers}
}
\value{
A set of GeoTIFF separate images
}
\description{
Save a RasterBrick into individual layers files GeoTIFFs
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterItanhanga.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)

# blocks saved in folder
lucC_blocks_raster_create(raster_obj = rb_class, number_blocks_xy = 2, save_images = TRUE)

lucC_blocks_raster_merge(path_open_GeoTIFFs = paste0(getwd(), "/Blocks_RasterBrick", sep = ""),
                         number_raster = 4, pattern_name = "Raster_Block_", is.rasterBrick = TRUE)

lucC_save_rasterBrick_by_layers(
path_name_GeoTIFF_Brick = paste0(getwd(),"/Blocks_RasterBrick/Mosaic_Raster_Block_.tif", sep = ""))



}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
