% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_treat_columns.R
\docType{data}
\name{lucC_select_columns}
\alias{lucC_select_columns}
\title{Select columns of a data set}
\usage{
lucC_select_columns(data_mtx = NULL, name_columns = NULL)
}
\arguments{
\item{data_mtx}{Matrix. A matrix with values obtained from predicates RECUR, EVOLVE, CONVERT or HOLDS}

\item{name_columns}{Character. Name of columns to remove from data set}
}
\value{
Matrix with without selected columns
}
\description{
Remove columns from a data_mtx
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterSample.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)
my_label <- c("Degradation", "Fallow_Cotton", "Forest", "Pasture", "Soy_Corn", "Soy_Cotton",
              "Soy_Fallow", "Soy_Millet", "Soy_Sunflower", "Sugarcane", "Urban_Area", "Water")
my_timeline <- c("2001-09-01", "2002-09-01", "2003-09-01", "2004-09-01", "2005-09-01",
                 "2006-09-01", "2007-09-01", "2008-09-01", "2009-09-01", "2010-09-01",
                 "2011-09-01", "2012-09-01", "2013-09-01", "2014-09-01", "2015-09-01",
                 "2016-09-01")

a <- lucC_pred_holds(raster_obj = rb_class, raster_class = "Forest",
                     time_interval = c("2001-09-01","2008-09-01"),
                     relation_interval = "contains", label = my_label,
                     timeline = my_timeline)
head(a)

# remove columns
lucC_select_columns(data_mtx = a, name_columns = c("2003-09-01"))


}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
