% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lucC_create_blocks.R
\docType{data}
\name{lucC_blocks_raster_create}
\alias{lucC_blocks_raster_create}
\title{Create blocks from RasterBrick in accordance with number of cells}
\usage{
lucC_blocks_raster_create (raster_obj = NULL, number_blocks_xy = 6, save_images = TRUE)
}
\arguments{
\item{raster_obj}{Raster. A raster stack with classified images}

\item{number_blocks_xy}{Numeric. A number of peaces to split raster, consider 2 peaces the raster will be cropped into 4 blocks. Default is 6}

\item{save_images}{Boolean. Write raster in file. Default is TRUE}
}
\value{
List with all the pieces in case you want to keep them in the memory.
}
\description{
Provide a data.frame with spatially aggregates the original raster it turns each aggregated cell into a polygon then the extent of each polygon is used to crop the original raster. \url{https://stackoverflow.com/questions/29784829/r-raster-package-split-image-into-multiples}
}
\examples{
\dontrun{
library(lucCalculus)

file <- c(system.file("extdata/raster/rasterItanhanga.tif", package = "lucCalculus"))
rb_class <- raster::brick(file)

# blocks saved in folder
lucC_blocks_raster_create(raster_obj = rb_class, number_blocks_xy = 2, save_images = TRUE)


}

}
\author{
Adeline M. Maciel
}
\keyword{datasets}
